SoundLoader fixed. Spiller av sfx og musikk. Oppdaterte musikkene i forkjellige states.   AGINC-72

+++ Vietnamese_Space_Program/Source/States/stateHighscore.cpp
@@ -136,7 +136,7 @@ void StateHighscore::destroy(sf::RenderWindow *window) {
     delete this->Texturetext[0].Player;
     delete this->PictureButtons[0];
 
-    for(int i = 0; i < sizeof(Texturetext) / sizeof(*Texturetext);i++)
+    for(int i = 0; (unsigned)i < sizeof(Texturetext) / sizeof(*Texturetext);i++)
     {
         delete this->Texturetext[i].Number;
     }
